refactor(featured products section): Remove shared folder
