+++ frontend/models/components/algolia-product-hit.ts
