+++ frontend/models/components/company.ts
