+++ frontend/models/components/product-preview.ts
