+++ frontend/models/page/index.ts
@@ -1,6 +1,6 @@
-import { FeaturedProductsSectionSchema } from '@models/shared/sections/featured-products-section';
-import { IFixitStatsSectionSchema } from '@models/shared/sections/ifixit-stats-section';
-import { SplitWithImageSectionSchema } from '@models/shared/sections/split-with-image-section';
+import { FeaturedProductsSectionSchema } from '@models/sections/featured-products-section';
+import { IFixitStatsSectionSchema } from '@models/sections/ifixit-stats-section';
+import { SplitWithImageSectionSchema } from '@models/sections/split-with-image-section';
 import { z } from 'zod';
 import { BrowseSectionSchema } from './sections/browse-section';
 import { HeroSectionSchema } from './sections/hero-section';
