+++ frontend/models/sections/featured-products-section.ts
