+++ frontend/models/sections/split-with-image-section.ts
